Implemented a work-around to a lumpy segfault and fixed the version command to only print the version.
The latest version of lumpy's included pairend_distro.py script sets mean and stdev to "NA" if there's not a lot of data (fewer than 1000 "elements"). While this is a reasonable step to take, the elimination of the data is not properly handled downstream. The construction of the -pe parameters to lumpy (in lumpyexpress) catch the instances of "NA" and omit the associated discordants files and settings for that file, but lumpy segfaults when this is done, probably because the number of split read files and discordant files is not the same. This work-around simply edits the variable in the distro script from 1000 to 1. The script is copied to the working directory, along with lumpy config and the lumpyexpress script.

Files checked in: tools/lumpyexpress/lumpy_wrapper.tcsh tools/lumpyexpress/lumpyexpress.xml